backend/requirements/base.txt
# This file was autogenerated by uv via the following command:
#    ./bin/compile_dependencies.sh
annotated-types==0.7.0
    # via pydantic
ape-pie==0.2.0
    # via zgw-consumers
asgiref==3.8.1
    # via
    #   django
    #   django-axes
asn1crypto==1.5.1
    # via webauthn
attrs==25.3.0
    # via
    #   glom
    #   jsonschema
    #   referencing
boltons==25.0.0
    # via
    #   face
    #   glom
cbor2==5.6.5
    # via webauthn
certifi==2025.1.31
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via cryptography
chardet==5.2.0
    # via prance
charset-normalizer==3.4.1
    # via requests
cryptography==44.0.2
    # via
    #   django-simple-certmanager
    #   josepy
    #   mozilla-django-oidc
    #   pyopenssl
    #   webauthn
django==5.2
    # via
    #   -r requirements/base.in
    #   django-admin-index
    #   django-axes
    #   django-formtools
    #   django-hijack
    #   django-jsonform
    #   django-otp
    #   django-phonenumber-field
    #   django-privates
    #   django-redis
    #   django-relativedelta
    #   django-rosetta
    #   django-sendfile2
    #   django-setup-configuration
    #   django-simple-certmanager
    #   django-solo
    #   django-two-factor-auth
    #   djangorestframework
    #   drf-spectacular
    #   maykin-2fa
    #   mozilla-django-oidc
    #   mozilla-django-oidc-db
    #   zgw-consumers
django-admin-index==4.0.0
    # via -r requirements/base.in
django-axes==7.0.2
    # via -r requirements/base.in
django-formtools==2.5.1
    # via django-two-factor-auth
django-hijack==3.7.1
    # via -r requirements/base.in
django-ipware==7.0.1
    # via django-axes
django-jsonform==2.23.2
    # via mozilla-django-oidc-db
django-ordered-model==3.7.4
    # via django-admin-index
django-otp==1.6.0
    # via django-two-factor-auth
django-phonenumber-field==8.0.0
    # via django-two-factor-auth
django-privates==3.1.0
    # via django-simple-certmanager
django-redis==5.4.0
    # via -r requirements/base.in
django-relativedelta==2.0.0
    # via zgw-consumers
django-rosetta==0.10.2
    # via -r requirements/base.in
django-sendfile2==0.7.2
    # via django-privates
django-setup-configuration==0.7.2
    # via zgw-consumers
django-simple-certmanager==2.4.1
    # via zgw-consumers
django-solo==2.4.0
    # via
    #   mozilla-django-oidc-db
    #   zgw-consumers
django-two-factor-auth==1.17.0
    # via maykin-2fa
djangorestframework==3.16.0
    # via
    #   -r requirements/base.in
    #   drf-spectacular
djangorestframework-camel-case==1.4.2
    # via -r requirements/base.in
docutils==0.21.2
    # via django-setup-configuration
drf-spectacular==0.28.0
    # via -r requirements/base.in
ecs-logging==2.2.0
    # via elastic-apm
elastic-apm==6.23.0
    # via -r requirements/base.in
face==24.0.0
    # via glom
furl==2.1.4
    # via ape-pie
glom==24.11.0
    # via mozilla-django-oidc-db
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
josepy==2.0.0
    # via mozilla-django-oidc
jsonschema==4.23.0
    # via
    #   drf-spectacular
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.4
    # via openapi-spec-validator
jsonschema-specifications==2024.10.1
    # via
    #   jsonschema
    #   openapi-schema-validator
lazy-object-proxy==1.11.0
    # via openapi-spec-validator
maykin-2fa==1.0.2
    # via -r requirements/base.in
mozilla-django-oidc==4.0.1
    # via mozilla-django-oidc-db
mozilla-django-oidc-db==0.22.0
    # via -r requirements/base.in
openapi-schema-validator==0.6.3
    # via openapi-spec-validator
openapi-spec-validator==0.7.1
    # via openapi3-parser
openapi3-parser==1.1.21
    # via -r requirements/base.in
orderedmultidict==1.0.1
    # via furl
packaging==25.0
    # via prance
pathable==0.4.4
    # via jsonschema-path
phonenumberslite==8.13.55
    # via django-two-factor-auth
pillow==11.1.0
    # via -r requirements/base.in
polib==1.2.0
    # via django-rosetta
prance==25.4.8.0
    # via openapi3-parser
psycopg2==2.9.10
    # via -r requirements/base.in
pycparser==2.22
    # via cffi
pydantic==2.11.3
    # via
    #   django-setup-configuration
    #   pydantic-settings
pydantic-core==2.33.1
    # via pydantic
pydantic-settings==2.9.1
    # via django-setup-configuration
pyjwt==2.10.1
    # via zgw-consumers
pyopenssl==25.0.0
    # via webauthn
pypng==0.20220715.0
    # via qrcode
python-dateutil==2.9.0.post0
    # via django-relativedelta
python-decouple==3.8
    # via -r requirements/base.in
python-dotenv==1.1.0
    # via
    #   -r requirements/base.in
    #   pydantic-settings
python-ipware==3.0.0
    # via django-ipware
pyyaml==6.0.2
    # via
    #   drf-spectacular
    #   jsonschema-path
    #   pydantic-settings
qrcode==7.4.2
    # via django-two-factor-auth
redis==5.2.1
    # via django-redis
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   ape-pie
    #   django-rosetta
    #   jsonschema-path
    #   mozilla-django-oidc
    #   prance
    #   zgw-consumers
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.10
    # via
    #   django-setup-configuration
    #   prance
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
sentry-sdk==2.25.1
    # via -r requirements/base.in
six==1.17.0
    # via
    #   furl
    #   orderedmultidict
    #   python-dateutil
    #   rfc3339-validator
sqlparse==0.5.3
    # via django
typing-extensions==4.13.1
    # via
    #   mozilla-django-oidc-db
    #   pydantic
    #   pydantic-core
    #   pyopenssl
    #   qrcode
    #   referencing
    #   typing-inspection
    #   zgw-consumers
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.3.0
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
uwsgi==2.0.28
    # via -r requirements/base.in
webauthn==2.5.2
    # via django-two-factor-auth
wrapt==1.17.2
    # via elastic-apm
zgw-consumers==0.38.0
    # via -r requirements/base.in
